Добавил валидацию данных

+++ Repository/UserRepository/UserRepository.cs
@@ -3,6 +3,9 @@ using marketplace_api.Models;
 using Microsoft.AspNetCore.JsonPatch;
 using marketplace_api.CustomExeption;
 using Microsoft.EntityFrameworkCore;
+using System.IdentityModel.Tokens.Jwt;
+using System.Security.Claims;
+using marketplace_api.Services.AuthService;
 
 namespace marketplace_api.Repository.UserRepository;
 
@@ -10,11 +13,13 @@ public class UserRepository : IUserRepository
 {
     private readonly AppDbContext _context;
     private readonly ILogger<AppDbContext> _logger;
+    private readonly JwtService _jwtService;
 
-    public UserRepository(AppDbContext context,ILogger<AppDbContext> logger)
+    public UserRepository(AppDbContext context,ILogger<AppDbContext> logger,JwtService jwtService)
     {
         _context = context;
         _logger = logger;
+        _jwtService = jwtService;
     }
 
     public async Task<User> CreateUserAsync(User user)
@@ -36,13 +41,14 @@ public class UserRepository : IUserRepository
     public async Task DeleteUserAsync(int id)
     {
         var result = await _context.Users.FindAsync(id); 
+        if (result == null)
+        {
+            throw new UserAlreadyExistsException("Данный пользователь уже существует");
+        }
+
         _context.Users.Remove(result);
+        await _context.SaveChangesAsync();
 
-    }
-
-    public Task DeleteUserAsync()
-    {
-        throw new NotImplementedException();
     }
 
     public async Task<List<User>> GetAllUsersAsync()
@@ -50,24 +56,30 @@ public class UserRepository : IUserRepository
         return _context.Users.ToList();
     }
 
-    public Task<User> GetByIndexUserAsync(int id)
+    public async Task<User> GetByIndexUserAsync(int id)
     {
-        throw new NotImplementedException();
+        var user = await _context.Users.FindAsync(id);
+
+        if (user == null)
+        {
+            throw new NotFoundExeption($"Пользователь с id {id} не найден");
+        }
+
+        return user;
     }
 
-    public Task<User> GetUserAsync()
+    public async Task<User> PatchUserAsync(JsonPatchDocument<User> user,int Id)
     {
-        throw new NotImplementedException();
-    }
+        var userUpdate = await _context.Users.FindAsync(Id);
 
-    public Task<JsonPatchDocument<User>> PatchUserAsync(JsonPatchDocument<User> user)
-    {
-        throw new NotImplementedException();
-    }
+        if(userUpdate == null)
+        {
+            throw new NotFoundExeption("данный пользователь не существкет");
+        }
+        user.ApplyTo(userUpdate);
+        await _context.SaveChangesAsync();
 
-    public Task<User> UpdateUserAsync(User user)
-    {
-        throw new NotImplementedException();
+        return userUpdate;
     }
 
     public async Task<User> GetUserByNameAsync(string name)
@@ -80,5 +92,21 @@ public class UserRepository : IUserRepository
         }
 
         return user;
+    }
+
+    public async Task<User> UpdateUserAsync(User user, int Id)
+    {
+        var userUpdate = await _context.Users.FindAsync(Id);
+        if (userUpdate == null)
+        {
+            throw new NotFoundExeption("Данный пользователь не сушествует");
+        }
+
+        userUpdate.Name = user.Name;
+        userUpdate.Email = user.Email;
+        userUpdate.HashPassword = user.HashPassword;
+        await _context.SaveChangesAsync();
+
+        return userUpdate;
     }
 }
